% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acv_fft.R
\name{acv_fft}
\alias{acv_fft}
\title{Fast auto-covariance with fftw}
\usage{
acv_fft(data_vector)
}
\arguments{
\item{data_vector}{a vector of centered data}
}
\value{
a vector of auto-covariance
}
\description{
Fast auto-covariance with fftw
}
