% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha.R
\name{alpha_plot}
\alias{alpha_plot}
\title{Internal companion function to \code{get_alpha}.}
\usage{
alpha_plot(df, df_bin, change_point, xdecades = 3, ydecades = 3)
}
\arguments{
\item{df}{a \code{data.frame} with the non-binned data}

\item{df_bin}{a \code{data.frame} with the binned data}

\item{change_point}{\code{numeric}, the change point identified by \code{get_alpha}}

\item{xdecades}{\code{numeric}, how decades should be plotted in the x-direction}

\item{ydecades}{\code{numeric}, how decades should be plotted in the y-direction}
}
\value{
a \code{ggplot} object
}
\description{
Internal companion function to \code{get_alpha}.
}
\keyword{zeta}
