% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_raster.R
\name{bin}
\alias{bin}
\title{Bins a power spectrum}
\usage{
bin(x, y, nbin)
}
\arguments{
\item{x}{\code{numeric}}

\item{y}{\code{numeric}}

\item{nbin}{\code{numeric}, the number of bins}
}
\value{
a matrix with with binned values and summary statistics
}
\description{
This code is (largely) a port from Matlab to \code{R} of the code from \href{https://web.mit.edu/perron/www/files/2DSpecTools-v1.1.zip}{2DSpecTools}: A 2D spectral analysis toolkit for Matlab by Taylor Perron et al.
}
\keyword{rotate_raster}
