% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_clamp.R
\name{clamp_raster_sigmas}
\alias{clamp_raster_sigmas}
\title{Filter outliers from a list of raster data using data from all rasters}
\usage{
clamp_raster_sigmas(
  raster_list,
  n_sigma = 3,
  band_id = 1,
  lower = TRUE,
  upper = TRUE,
  lower_clamp = -Inf,
  upper_clamp = Inf,
  use_values = FALSE
)
}
\arguments{
\item{raster_list}{a \code{list} of \code{stars} objects}

\item{n_sigma}{\code{numeric}, number of standard deviations to consider to calculate the percentile corresponding to values to filter, default to 3}

\item{band_id}{\code{numeric}, identifier of the band of the \code{stars} object}

\item{lower}{\code{logical}, if \code{TRUE} a lower clamp value is computed using the specified number of standard deviations}

\item{upper}{\code{logical}, if \code{TRUE} an upper clamp value is computed using the specified number of standard deviations}

\item{lower_clamp}{\code{numeric}, a value for the lower clamp, default to \code{-Inf}}

\item{upper_clamp}{\code{numeric}, a value for the upper clamp, default to \code{Inf}}

\item{use_values}{logical. If \code{FALSE} values outside the clamping range become \code{NA}, if \code{TRUE}, they get the extreme values, default to \code{TRUE}}
}
\value{
a named \code{list} of with the clamped \code{stars} objects and the clamped values of all the rasters (from ploting purposes)
}
\description{
This function uses a \code{list} as input so that the object can have vastly different resolutions.
}
