% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_Hurst_rasters.R
\name{compute_Hurst_rasters_internal}
\alias{compute_Hurst_rasters_internal}
\title{Compute Hurst rasters from a log-log regression between scales and mean standard deviation}
\usage{
compute_Hurst_rasters_internal(
  mean_sd_rasters,
  R,
  mean_res,
  L_raster,
  min_res = 1.87
)
}
\arguments{
\item{mean_sd_rasters}{list of rasters of mean standard deviation computed at R scales}

\item{R}{list of scales (aggregation factors)}

\item{mean_res}{numeric, resolution to compute scales in meters}

\item{L_raster}{target raster at scale L}

\item{min_res}{minimum resolution in elevation, 1.87 m from Haneberg, 2006}
}
\value{
a RasterStack with four layers q, H, p-value and R squared values
}
\description{
Compute Hurst rasters from a log-log regression between scales and mean standard deviation
}
\keyword{Hurst}
