% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_zeta.R
\name{correct_zeta}
\alias{correct_zeta}
\title{Correct zeta}
\usage{
correct_zeta(raster_list, target_id)
}
\arguments{
\item{raster_list}{a \code{list} of \code{stars} objects}

\item{target_id}{the id of the layer to be masked}
}
\value{
a \code{list} of \code{stars}
}
\description{
Correct zeta
}
