% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossscale_correlations.R
\name{crossscale_correlations}
\alias{crossscale_correlations}
\title{Creates a facet plot of correlation between \code{selected} attributes from a \code{raster_list} representing different \code{spatial_scales}}
\usage{
crossscale_correlations(
  raster_list = NULL,
  selected = NULL,
  spatial_scales = NULL,
  corr_type = "spearman",
  clamp_raster = FALSE
)
}
\arguments{
\item{raster_list}{a \code{list} of \code{stars} or \code{RasterStack} objects}

\item{selected}{\code{character}, the selected attributes}

\item{spatial_scales}{\code{numeric}, the spatial scales corresponding to each raster in \code{raster_list}}

\item{corr_type}{\code{character}, the type of correlation to compute, 'pearson' or 'spearman'}

\item{clamp_raster}{\code{logical}, if \code{TRUE} clamps the rasters using \code{clamp_params}}
}
\value{
if \code{clamp} is \code{FALSE} a \code{ggplot} object, otherwise a named \code{list} with two elements, a \code{ggplot} object and a list of clamped rasters.
}
\description{
Creates a facet plot of correlation between \code{selected} attributes from a \code{raster_list} representing different \code{spatial_scales}
}
\keyword{postprocessing}
