% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft2d.R
\name{detrend_dem}
\alias{detrend_dem}
\title{Detrends a DEM with \code{pracma::odregress()}}
\usage{
detrend_dem(dem)
}
\arguments{
\item{dem}{a \code{RasterLayer}}
}
\value{
the detrended raster as a \code{RasterLayer}
}
\description{
Detrends a DEM with \code{pracma::odregress()}
}
\keyword{fft2d}
