% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft2d.R
\name{fft2D}
\alias{fft2D}
\title{Perfoms a two-dimensional Fourier transform}
\usage{
fft2D(M, dx, dy, Hann = TRUE)
}
\arguments{
\item{M}{a \code{matrix}}

\item{dx}{the spacing in the x direction in map units (usually meters)}

\item{dy}{the spacing in the y direction in map units (usually meters)}

\item{Hann}{\code{logical}, if \code{TRUE} performs a Hann windowing, default to \code{FALSE}}
}
\value{
a list with 4 elements: \code{spectral_power_matrix}, \code{radial_frequency_matrix}, \code{spectral_power_vector} and \code{radial_frequency_vector}.
}
\description{
This code is (largely) a port from Matlab to \code{R} of the code from \href{https://web.mit.edu/perron/www/files/2DSpecTools-v1.1.zip}{2DSpecTools}: A 2D spectral analysis toolkit for Matlab by Taylor Perron et al.
As \code{dx} and \code{dy} refers to spacing in map units, it is recommended, if applicable, that the original raster is projected from latitude, longitude to a planar projection.
}
\keyword{fft2d}
