% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_raster.R
\name{filter_spectral_power_matrix}
\alias{filter_spectral_power_matrix}
\title{Returns a quantile-filtered raster of the logarithm of the real-part of the normalized spectral power matrix}
\usage{
filter_spectral_power_matrix(
  normalized_spectral_power_matrix,
  FT2D,
  quantile_prob = c(0.99)
)
}
\arguments{
\item{normalized_spectral_power_matrix}{\code{matrix}, normalized spectral power matrix}

\item{FT2D}{\code{list}, results from \code{fft2d()} with objects radial frequency and spectral power matrices}

\item{quantile_prob}{vector of quantile probabilities, default to \code{c(0.99)}}
}
\value{
a \code{RasterLayer} of the normalized spectral power matrix filtered for components above \code{quantile_prob}
}
\description{
Returns a quantile-filtered raster of the logarithm of the real-part of the normalized spectral power matrix
}
\keyword{rotate_raster}
