% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_par_perp.R
\name{find_par_perp}
\alias{find_par_perp}
\title{Add parallel and perpendicular directions}
\usage{
find_par_perp(raster_list, xdir_id, ydir_id)
}
\arguments{
\item{raster_list}{a \code{list} of \code{stars} objects}

\item{xdir_id}{the id of the layer with the x direction}

\item{ydir_id}{the id of the layer with the y direction}
}
\value{
a \code{list} of \code{stars}
}
\description{
Add parallel and perpendicular directions
}
