% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_values_check.R
\name{four_values_check}
\alias{four_values_check}
\title{Check that at least four values of the raster stack are not \code{NA} for each element of \code{raster_list}}
\usage{
four_values_check(raster_list)
}
\arguments{
\item{raster_list}{a \code{list} of \code{stars} objects}
}
\value{
\code{logical}
}
\description{
Check that at least four values of the raster stack are not \code{NA} for each element of \code{raster_list}
}
