% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha.R
\name{get_all_alpha}
\alias{get_all_alpha}
\title{Wrapper to extract all roughness exponents from the HHCFs}
\usage{
get_all_alpha(hhcf, dr)
}
\arguments{
\item{hhcf}{\code{list} from \code{get_hhcf()}}

\item{dr}{\code{numeric}, spacing of the values along the axis}
}
\value{
a \code{data.frame}
}
\description{
Wrapper to extract all roughness exponents from the HHCFs
}
\keyword{zeta}
