% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{get_all_R_L}
\alias{get_all_R_L}
\title{Get a list of upper spatial scale at which the roughness coefficient is calculated}
\usage{
get_all_R_L(Lmax, factor_limit, only = NULL, logfilter = TRUE, len = 32)
}
\arguments{
\item{Lmax}{maximum value of L}

\item{factor_limit}{a threshold of the minimum length of factors for a scale to be accepted}

\item{only}{default to \code{NULL}, if \code{only == 610} only factors of 6 and 10 are considered for a quick search, if \code{only == 12} only factors of 12 are considered for the fastest search}

\item{logfilter}{logical, default to \code{TRUE}, if \code{TRUE} the returned scale are logaritmically spaced}

\item{len}{integer, number of scales returned, default to 32, len only works if \code{logfilter == TRUE}}
}
\value{
a list with two elements \code{allRs} and \code{allL}
}
\description{
Get a list of upper spatial scale at which the roughness coefficient is calculated
}
