% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha.R
\name{get_alpha_}
\alias{get_alpha_}
\title{Estimates the roughness exponent from the HHCF from spline fit.}
\usage{
get_alpha_(row, dr, xi, do_plot = FALSE)
}
\arguments{
\item{row}{array of hhcf values}

\item{dr}{\code{numeric}, spacing of the values along the axis}

\item{xi}{\code{numeric}, auto-correlation length}

\item{do_plot}{\code{logical}, plot the fit, default to \code{FALSE}}
}
\description{
This function will return \code{NA} if the HHFC is monotically increasing.
}
\keyword{zeta}
