% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beta.R
\name{get_beta}
\alias{get_beta}
\title{Derive the slope of the binned power spectrum}
\usage{
get_beta(binned_power_spectrum, FT2D, do_plot = FALSE)
}
\arguments{
\item{binned_power_spectrum}{\code{matrix}, a binned spectrum from \code{bin()}}

\item{FT2D}{a \code{list} from \code{fft2d()}}

\item{do_plot}{\code{logical}, plot the fit, default to \code{FALSE}}
}
\value{
a \code{data.frame}
}
\description{
Derive the slope of the binned power spectrum
}
\keyword{fft2d}
