% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{get_factors}
\alias{get_factors}
\title{Get factor of an integer}
\usage{
get_factors(x)
}
\arguments{
\item{x, }{numeric, converted to integer}
}
\value{
list of factors
}
\description{
Get factor of an integer
}
