% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_raster.R
\name{get_fourier_angle}
\alias{get_fourier_angle}
\title{Get the angle of the 2D Fourier spectrum}
\usage{
get_fourier_angle(Pmn_truncated)
}
\arguments{
\item{Pmn_truncated}{truncated normalized power spectrum}
}
\value{
angle of the main component of the Fourier spectrum
}
\description{
Get the angle of the 2D Fourier spectrum
}
\keyword{rotate_raster}
