% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hhcf.R
\name{get_hhcf}
\alias{get_hhcf}
\title{This wrapper function derives the one-dimensional \code{HHCF()} for all columns or rows of a matrix}
\usage{
get_hhcf(mat, margin = 1, detrend = FALSE, get_autocorr = FALSE)
}
\arguments{
\item{mat}{a \code{matrix}}

\item{margin}{indicate the direction, 1 = row, 2 = column, passed to \code{apply}'s \code{MARGIN}}

\item{detrend}{\code{logical} indicates if the data need to be detrended, default to \code{FALSE}}

\item{get_autocorr}{\code{logical}, indicaites if the auto-correlation length is calculated}
}
\description{
This wrapper function derives the one-dimensional \code{HHCF()} for all columns or rows of a matrix
}
\keyword{zeta}
