% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hhcf.R
\name{get_hhcf_}
\alias{get_hhcf_}
\title{This wrapper function derives the one-dimensional \code{HHCF()} for all columns or rows of a matrix}
\usage{
get_hhcf_(mat, dr, margin = 1, limlen = 30, average = FALSE)
}
\arguments{
\item{mat}{a \code{matrix}}

\item{dr}{\code{numeric}, spacing of the values along the axis}

\item{margin}{indicate the direction, 1 = row, 2 = column, passed to \code{apply}'s \code{MARGIN}}

\item{limlen}{\code{numeric} the minimum length of data}
}
\description{
This wrapper function derives the one-dimensional \code{HHCF()} for all columns or rows of a matrix
}
\keyword{zeta}
