% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zeta.R
\name{get_kruskal_flag}
\alias{get_kruskal_flag}
\title{Computes a Kruskal-Wallis ranksum test for significant differences between the distributions of alpha}
\usage{
get_kruskal_flag(alpha_x, alpha_y, var)
}
\arguments{
\item{alpha_x}{\code{data.frame} from get_all_alpha_()`}

\item{alpha_y}{\code{data.frame} from get_all_alpha_()`}

\item{var}{\code{character} the variable to compare}
}
\value{
a \code{logical} value: \code{TRUE} if the distribution are statistically different, \code{FALSE} otherwise
}
\description{
Computes a Kruskal-Wallis ranksum test for significant differences between the distributions of alpha
}
