% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_res.R
\name{get_mean_res}
\alias{get_mean_res}
\title{Compute the mean resolution of lat long raster in meters}
\usage{
get_mean_res(r, L, crs.ref = crs_ref)
}
\arguments{
\item{r}{aggegated raster}

\item{L}{aggregation factor}

\item{crs.ref}{reference crs}
}
\value{
numeric, mean resolution in m
}
\description{
Compute the mean resolution of lat long raster in meters
}
\keyword{Hurst}
