% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mean_sd_rasters.R
\name{get_mean_sd_rasters}
\alias{get_mean_sd_rasters}
\title{Compute box-counting mean standard deviation rasters}
\usage{
get_mean_sd_rasters(dem, L, R)
}
\arguments{
\item{dem}{initial DEM}

\item{L}{integer}

\item{R}{vector of values}
}
\value{
list of rasters
}
\description{
Compute box-counting mean standard deviation rasters
}
\keyword{Hurst}
