% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_raster.R
\name{get_normalized_spectral_power_matrix}
\alias{get_normalized_spectral_power_matrix}
\title{Normalizes the spectral power matrix with the background spectrum}
\usage{
get_normalized_spectral_power_matrix(B, FT2D)
}
\arguments{
\item{B}{\code{matrix} from \code{bin()}, binned radial frequency and spectral power vectors}

\item{FT2D}{\code{list}, results from \code{fft2d()} with objects radial frequency and spectral power matrices}
}
\value{
\code{matrix}, normalized spectral power matrix
}
\description{
Normalizes the spectral power matrix with the background spectrum
}
\keyword{rotate_raster}
