% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_radial_angle.R
\name{get_radial_angle}
\alias{get_radial_angle}
\title{Wrapper function to derive the anisotropy exponent from a non-rotated raster}
\usage{
get_radial_angle(rstr, raster_resolution, angle_step, niter)
}
\arguments{
\item{rstr}{\code{Raster} object}

\item{raster_resolution}{\code{numeric} the resolution of \code{rstr} in meters}

\item{angle_step}{\code{numeric}, angular step}

\item{niter}{\code{numeric}, number of random rotations}
}
\value{
a \code{data.frame}
}
\description{
Wrapper function to derive the anisotropy exponent from a non-rotated raster
}
\keyword{zeta}
