% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zeta.R
\name{get_zeta}
\alias{get_zeta}
\title{Wrapper function to derive the anisotropy exponent from a non-rotated raster}
\usage{
get_zeta(
  rstr,
  raster_resolution,
  nbin = 20,
  .Hann = TRUE,
  .quantile_prob = c(0.9999),
  .prob = 0.999,
  full = FALSE
)
}
\arguments{
\item{rstr}{\code{Raster} object}

\item{raster_resolution}{\code{numeric} the resolution of \code{rstr} in meters}

\item{nbin}{\code{numeric}, the number of bins}

\item{.Hann}{\code{logical}, if \code{TRUE} performs a Hann windowing, default to \code{TRUE}, passed to \code{fft2D()}}

\item{.quantile_prob}{vector of quantile probabilities, default to \code{c(0.9999)}, passed to \code{filter_spectral_power_matrix()}}

\item{.prob, }{\code{numeric} the value of the probabilities passed to \code{quantile()}, passed to \code{filter_alpha()}}

\item{full, }{\code{logical} if \code{TRUE} all results are returned, else, the default, results of interest are returned}
}
\value{
a \code{data.frame}
}
\description{
Wrapper function to derive the anisotropy exponent from a non-rotated raster
}
\keyword{zeta}
