% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zeta.R
\name{get_zeta_}
\alias{get_zeta_}
\title{Internal function to derive the anisotropy exponent from two roughness exponents}
\usage{
get_zeta_(alpha_1, alpha_2, IQR_1, IQR_2, kruskal_flag)
}
\arguments{
\item{alpha_1}{\code{numeric}, a roughness coefficient}

\item{alpha_2}{\code{numeric}, a roughness coefficient}

\item{IQR_1}{\code{numeric}, an IQR}

\item{IQR_2}{\code{numeric}, an IQR}

\item{kruskal_flag}{\code{logical} flag returned by \code{get_kruskal_flag()}}
}
\value{
a \code{numeric} anisotropy exponent
}
\description{
If any IQR is higher than 0.225, this function returns \code{NA}. The threshold value of 0.225 is the IQR of a normal distribution with a standard deviation equals to 0.5/3, a three-sigma range between 0 and 1.
}
\keyword{zeta}
