% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zeta.R
\name{get_zeta_radial}
\alias{get_zeta_radial}
\title{Wrapper function to derive the anisotropy exponent from a non-rotated raster}
\usage{
get_zeta_radial(rstr, raster_resolution, angle_step, niter, full)
}
\arguments{
\item{rstr}{\code{Raster} object}

\item{raster_resolution}{\code{numeric} the resolution of \code{rstr} in meters}

\item{angle_step}{\code{numeric}, angular step}

\item{niter}{\code{numeric}, number of random rotations}

\item{full, }{\code{logical} if \code{TRUE} all results are returned, else, the default, results of interest are returned}
}
\value{
a \code{data.frame}
}
\description{
Wrapper function to derive the anisotropy exponent from a non-rotated raster
}
\keyword{zeta}
