% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zeta_raster.R
\name{get_zeta_raster}
\alias{get_zeta_raster}
\title{Get the value of the anisotropy exponent for a tile of elevation}
\usage{
get_zeta_raster(DEM, tiles, .zeta_df = NULL, ...)
}
\arguments{
\item{DEM}{an elevation raster}

\item{tiles}{a \code{RasterLayer} or a \code{SpatialPolygons} to tiles \code{DEM}}

\item{.zeta_df}{null or a data.frame with column zeta and a number of rows equals to the number of cells of tiles}

\item{...}{to pass the \code{raster_resolution} and \code{vertical_accuracy} argument to \code{get_zeta_df()}}
}
\value{
stack of raster of size tiles with anisotropy exponent values and a number of layers equals to the number of column in .zeta_df
}
\description{
Get the value of the anisotropy exponent for a tile of elevation
}
