% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft2d.R
\name{Hann2D}
\alias{Hann2D}
\title{A two-dimensional Hann window}
\usage{
Hann2D(M)
}
\arguments{
\item{M}{a \code{matrix}}
}
\value{
a list with two elements the windowed matrix \code{H} and the sum of the coefficients \code{Wss}
}
\description{
This code is (largely) a port from Matlab to \code{R} of the code from \href{https://web.mit.edu/perron/www/files/2DSpecTools-v1.1.zip}{2DSpecTools}: A 2D spectral analysis toolkit for Matlab by Taylor Perron et al.
}
\keyword{fft2d}
