% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hhcf.R
\name{HHCF}
\alias{HHCF}
\title{Internal; derives a one-dimensional height-height correlation function}
\usage{
HHCF(row)
}
\arguments{
\item{row}{array of values}
}
\value{
array with the values of the height-height correlation function
}
\description{
Internal; derives a one-dimensional height-height correlation function
}
\keyword{zeta}
