% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logtransform_layer.R
\name{logtransform_layer}
\alias{logtransform_layer}
\title{Apply a log transform to a given layer}
\usage{
logtransform_layer(raster_list, target_id)
}
\arguments{
\item{raster_list}{a \code{list} of \code{stars} objects}

\item{target_id}{the id of the layer to be masked}
}
\value{
a \code{list} of \code{stars}
}
\description{
Apply a log transform to a given layer
}
