% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_angular.R
\name{make_angular}
\alias{make_angular}
\title{Transform a 360 degrees variable to a 180 degrees}
\usage{
make_angular(raster_list, target_id, mode = "half")
}
\arguments{
\item{raster_list}{a \code{list} of \code{stars} objects}

\item{target_id}{the id of the layer to be masked}

\item{mode}{\code{character}, should the function consider a full or a half-circle}
}
\value{
a \code{list} of \code{stars}
}
\description{
Transform a 360 degrees variable to a 180 degrees
}
\keyword{postprocessing}
