% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_leaflet_map.R
\name{make_leaflet_map}
\alias{make_leaflet_map}
\title{Produces a leaflet map}
\usage{
make_leaflet_map(
  clamped,
  ttl = "values",
  n_class = 10,
  circular = FALSE,
  style = "continuous",
  groups
)
}
\arguments{
\item{clamped}{named \code{list} from \code{clamp_raster_sigmas()}}

\item{ttl}{\code{character}, title}

\item{n_class}{\code{numeric}, number of bins for the color scheme}

\item{circular}{\code{logical}, if \code{TRUE} the value is transformed so be between 0 and 180, default to \code{FALSE}}

\item{style}{\code{character}, changes between continuous and binned values for the colorscale, default to \code{continuous}}

\item{groups}{\code{character}, overlay groups to adjust map visuals}
}
\value{
a \code{leaflet} object
}
\description{
Produces a leaflet map
}
