% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stacked_density_plots.R
\name{make_stacked_density_plot}
\alias{make_stacked_density_plot}
\title{Makes stacked density plot}
\usage{
make_stacked_density_plot(
  raster_list,
  band_id = 1,
  var_name = "variable",
  limx = c(0, 1),
  logscale = FALSE,
  spatial_scales = NULL,
  ...
)
}
\arguments{
\item{raster_list}{a \code{list} of \code{stars} or \code{RasterStack} objects}

\item{band_id}{\code{numeric}, identifier of the band of the \code{stars} object}

\item{var_name}{\code{character}, used for the title of the graph}

\item{limx}{\code{numeric} limits of the x-axis, default to \code{c(0,1)}}

\item{logscale}{\code{logical} if \code{TRUE} the x-axis is log-transformed}

\item{spatial_scales}{\code{numeric}, the spatial scales corresponding to each raster in \code{raster_list}}

\item{...}{passed to \code{ggplot2::scale_color_viridis_c()}}
}
\value{
a \code{ggplot()} object
}
\description{
Makes stacked density plot
}
