% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mask_layer_from_layer.R
\name{mask_layer_from_layer}
\alias{mask_layer_from_layer}
\title{Mask a layer with another}
\usage{
mask_layer_from_layer(
  raster_list,
  target_id,
  mask_id,
  option = "pval",
  pval_threshold = 0.05,
  r2_threshold = 0.95
)
}
\arguments{
\item{raster_list}{a \code{list} of \code{stars} objects}

\item{target_id}{the id of the layer to be masked}

\item{mask_id}{the id of the layer to use as a mask, currently default to pvalue mask with a threshold at 0.05}

\item{option}{\code{character} one of \code{pval}, \code{r2}, \code{NA}}

\item{pval_threshold}{\code{numeric} threshold of p-values, default to 0.05}

\item{r2_threshold}{\code{numeric} threshold of r squared, default to 0.95}
}
\value{
a \code{list} of \code{stars}
}
\description{
Mask a layer with another
}
