% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft2d_visualizations.R
\name{matrix_for_rayshader}
\alias{matrix_for_rayshader}
\title{Transforms matrix for be used by \code{rayshader}.}
\usage{
matrix_for_rayshader(m)
}
\arguments{
\item{m}{a \code{matrix}}
}
\value{
a \code{matrix}
}
\description{
Transforms matrix for be used by \code{rayshader}.
}
\keyword{fft2d_viz}
