% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_Hurst_rasters.R
\name{merge_Hurst_rasters}
\alias{merge_Hurst_rasters}
\title{merge all Hurst raster}
\usage{
merge_Hurst_rasters(
  Hurst_dir = file.path("F:/hguillon/research/exploitation/out/run127")
)
}
\arguments{
\item{Hurst_dir}{a file.path to a directory containing tiled Hurst rasters}
}
\description{
merge all Hurst raster
}
