% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stacked_density_plots.R
\name{modes_from_stacked_density}
\alias{modes_from_stacked_density}
\title{Extract the modal values of the stacked density plot}
\usage{
modes_from_stacked_density(
  plot_object,
  spatial_scales = NULL,
  var_name = "variable",
  ...
)
}
\arguments{
\item{plot_object}{a \code{ggplot} object}

\item{spatial_scales}{\code{numeric}, the spatial scales corresponding to each raster in \code{raster_list}}

\item{var_name}{\code{character}, used for the title of the graph}

\item{...}{passed to \code{ggplot2::scale_color_viridis_c()}}
}
\value{
a \code{list} with three \code{ggplot} elements: a plot in the variable - density space, a plot in the variable mode - scale space, a density plot with the modes added.
}
\description{
Extract the modal values of the stacked density plot
}
