% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft2d.R
\name{pad_mat}
\alias{pad_mat}
\title{Pads a matrix with zeros so it has a given dimension}
\usage{
pad_mat(x, n, m)
}
\arguments{
\item{x}{a \code{matrix}}

\item{n}{the desired number of rows of \code{x}}

\item{m}{the desired number of columns of \code{x}}
}
\value{
a padded version of \code{x} with dimension \verb{n,m}
}
\description{
This function is useful to pad a matrix so that the dimension is a factor of 2.
}
\keyword{fft2d}
