% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raster_select.R
\name{raster_select}
\alias{raster_select}
\title{Selecting raster bands}
\usage{
raster_select(raster_list, band_id)
}
\arguments{
\item{raster_list}{a \code{list} of \code{stars} or \code{RasterStack} objects}

\item{band_id}{\code{numeric}, identifier of the band of the \code{stars} object}
}
\description{
Selecting raster bands
}
