% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_zeta_raster.R
\name{read_zeta_raster}
\alias{read_zeta_raster}
\title{Read zeta raster from destination}
\usage{
read_zeta_raster(
  out_dir = "/out/run128/out",
  Lmax = 10000,
  raster_resolution = 10
)
}
\arguments{
\item{out_dir}{\code{character} file path to folder}

\item{Lmax}{\code{numeric} maximum scale}

\item{raster_resolution}{\code{numeric} raster resolution, to determine the scales}
}
\value{
a \code{list} with two elements containing the rasters as a \code{list} of \code{stars} objects and the spatial scales
}
\description{
Read zeta raster from destination
}
