% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_spatial_noise.R
\name{reduce_spatial_noise}
\alias{reduce_spatial_noise}
\title{Convolves with a 3x3 mean kernel with weights of 1}
\usage{
reduce_spatial_noise(raster_list, .NAonly = FALSE)
}
\arguments{
\item{raster_list}{a \code{list} of \code{stars} or \code{RasterStack} objects}

\item{.NAonly}{\code{logical}. If \code{TRUE}, only cell values that are NA are replaced with the computed focal values}
}
\value{
a \code{list} of \code{stars} or \code{RasterStack} objects
}
\description{
Convolves with a 3x3 mean kernel with weights of 1
}
\keyword{postprocessing}
