% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_map.R
\name{save_map}
\alias{save_map}
\title{Creates and save a map of a given variablw}
\usage{
save_map(
  raster_list,
  groups,
  n_class = 10,
  ttl = "zeta",
  begin = 0.1,
  end = 0.95,
  direction = 1,
  option = "inferno",
  out_path = "out",
  col_style = NULL,
  cstyle = "equal"
)
}
\arguments{
\item{raster_list}{a named list with two elements \code{rasters} and \code{values} supplied from `raster_select()}

\item{groups}{the groups to use in the maps}

\item{n_class}{\code{numeric} number of classes}

\item{ttl}{\code{character} the title of the map}

\item{begin}{\code{numeric} passed to \code{tmap::tm_raster()}}

\item{end}{\code{numeric} passed to \code{tmap::tm_raster()}}

\item{direction}{\code{numeric} passed to \code{tmap::tm_raster()}}

\item{option}{\code{character} passed to \code{tmap::tm_raster()}}

\item{out_path}{\code{character} path to save the images}

\item{col_style}{\code{character}, default to \code{NULL}, passed as \code{style} to \code{tmap::tm_raster()}}

\item{cstyle}{\code{character}, default to \code{equal}, passed as \code{style} to \code{classInt::classIntervals()}}
}
\description{
Creates and save a map of a given variablw
}
