% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signifNA.R
\name{signifNA}
\alias{signifNA}
\title{Internal function to gracefully handles NA in signif}
\usage{
signifNA(value, digits = 3)
}
\arguments{
\item{digits}{\code{numeric} number of significant digits, default to 3}
}
\description{
Internal function to gracefully handles NA in signif
}
\keyword{internal}
