% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_clamp.R
\name{slice_clamp}
\alias{slice_clamp}
\title{Slicing and clamping of rasters}
\usage{
slice_clamp(
  raster_list = NULL,
  att_names = NULL,
  selected = NULL,
  clamp_raster = FALSE,
  clamp_params = NULL
)
}
\arguments{
\item{raster_list}{a \code{list} of \code{stars} or \code{RasterStack} objects}

\item{att_names}{\code{character}, the attributes corresponding to each band}

\item{selected}{\code{character}, the selected attributes}

\item{clamp_raster}{\code{logical}, if \code{TRUE} clamps the rasters using \code{clamp_params}}

\item{clamp_params}{\code{data.frame} of parameters to clamp rasters using \code{clamp_raster_sigmas()}}
}
\value{
if \code{clamp} is \code{FALSE} a \code{ggplot} object, otherwise a named \code{list} with two elements, a \code{ggplot} object and a list of clamped rasters.
}
\description{
Slicing and clamping of rasters
}
\keyword{postprocessing}
