% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft2d.R
\name{spectrum_plot}
\alias{spectrum_plot}
\title{Plot a radial Fourier spectrum}
\usage{
spectrum_plot(binned_power_spectrum, FT2D, xdecades = 3, ydecades = 3)
}
\arguments{
\item{binned_power_spectrum}{\code{matrix}, a binned spectrum from \code{bin()}}

\item{FT2D}{a \code{list} from \code{fft2d()}}
}
\value{
a \code{ggplot} object
}
\description{
Plot a radial Fourier spectrum
}
\keyword{fft2d}
