% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha.R
\name{summarise_alpha}
\alias{summarise_alpha}
\title{Wrapper to summarize \code{alpha} values}
\usage{
summarise_alpha(alpha)
}
\arguments{
\item{alpha, }{a \code{data.frame} returned by \code{get_all_alpha()} or \code{filter_alpha()}}
}
\value{
a summarized \code{data.frame}
}
\description{
Wrapper to summarize \code{alpha} values
}
\keyword{zeta}
