% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fft2d_visualizations.R
\name{view_matrix}
\alias{view_matrix}
\title{Visualizes values of a matrix}
\usage{
view_matrix(m, ply = TRUE)
}
\arguments{
\item{m}{a \code{matrix}}

\item{ply}{\code{logical}}
}
\value{
if \code{ply = TRUE} returns a \code{plotly::ggploty()} object, if \code{FALSE} returns a \code{ggplot} object
}
\description{
Visualizes values of a matrix
}
\keyword{fft2d_viz}
