% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha.R
\name{filter_alpha}
\alias{filter_alpha}
\title{Filters the slope values of alpha returned by \code{get_all_alpha()} between zero and the 99.9\% quantile of the initial slopes}
\usage{
filter_alpha(alpha, prob = 0.999)
}
\arguments{
\item{alpha, }{a \code{data.frame} returned by \code{get_all_alpha()}}

\item{prob, }{\code{numeric} the value of the probabilities passed to \code{quantile()}}
}
\value{
a \code{data.frame}
}
\description{
Filters the slope values of alpha returned by \code{get_all_alpha()} between zero and the 99.9\% quantile of the initial slopes
}
\keyword{zeta}
