% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_raster.R
\name{get_fourier_angle}
\alias{get_fourier_angle}
\title{Get the angle of the 2D Fourier spectrum}
\usage{
get_fourier_angle(filtered_spectral_power_matrix, FT2D, bandwidth = 5)
}
\arguments{
\item{filtered_spectral_power_matrix}{truncated normalized power spectrum}

\item{FT2D}{\code{list}, results from \code{fft2d()} with objects radial frequency and spectral power matrices}

\item{bandwidth}{\code{numeric} the bandwidth for the kernel smoothing of the circular density calculation}
}
\value{
angle of the main component of the Fourier spectrum, counted counter-clockwise
}
\description{
Get the angle of the 2D Fourier spectrum
}
\keyword{rotate_raster}
