% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zeta.R
\name{get_zeta_}
\alias{get_zeta_}
\title{Internal function to derive the anisotropy exponent from two roughness exponents}
\usage{
get_zeta_(alpha_1, alpha_2)
}
\arguments{
\item{alpha_1}{\code{numeric}, a roughness coefficient}

\item{alpha_2}{\code{numeric}, a roughness coefficient}
}
\value{
a \code{numeric} anisotropy exponent
}
\description{
Internal function to derive the anisotropy exponent from two roughness exponents
}
\keyword{zeta}
