% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_raster.R
\name{rotate_raster}
\alias{rotate_raster}
\title{Rotates the detrended DEM according to the main direction of the Fourier spectrum}
\usage{
rotate_raster(dem_detrended, dem, ang_fourier, mean_res)
}
\arguments{
\item{dem_detrended}{\code{RasterLayer}, detrended DEM}

\item{dem}{\code{RasterLayer}, original (cropped) DEM}

\item{ang_fourier}{\code{numeric}, angle of the main component of the 2D Fourier spectrum}
}
\value{
a \code{matrix} corresponding to the DEM rotated in the main direction of the Fourier spectrum
}
\description{
Rotates the detrended DEM according to the main direction of the Fourier spectrum
}
\keyword{rotate_raster}
