% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_raster.R
\name{rotate_raster}
\alias{rotate_raster}
\title{Rotates the detrended DEM according to the main direction of the Fourier spectrum}
\usage{
rotate_raster(mm, ang_fourier)
}
\arguments{
\item{mm}{a \code{RasterLayer}}

\item{ang_fourier}{\code{numeric}, angle of the main component of the 2D Fourier spectrum}
}
\value{
a \code{matrix} corresponding to the DEM rotated in the main direction of the Fourier spectrum
}
\description{
Rotates the detrended DEM according to the main direction of the Fourier spectrum
}
\keyword{rotate_raster}
